requirements.txt
Django==1.6.5
djangorestframework==2.3.13
requests==2.3.0
